Coriolis implemented for Python engines. Shot and ShotProps moved to own module.
