+++ py_ballisticcalc/engines/scipy_engine.py
@@ -63,7 +63,7 @@ from typing_extensions import List, Optional, Tuple, Union, override
 
 # Local imports
 from py_ballisticcalc._compat import bisect_left_key
-from py_ballisticcalc.conditions import ShotProps, Wind
+from py_ballisticcalc.conditions import Wind
 from py_ballisticcalc.engines.base_engine import (
     BaseEngineConfig,
     BaseEngineConfigDict,
@@ -73,6 +73,7 @@ from py_ballisticcalc.engines.base_engine import (
 )
 from py_ballisticcalc.exceptions import OutOfRangeError, RangeError, ZeroFindingError
 from py_ballisticcalc.logger import logger
+from py_ballisticcalc.shot import ShotProps
 from py_ballisticcalc.trajectory_data import HitResult, TrajFlag, TrajectoryData
 from py_ballisticcalc.unit import Angular, Distance
 from py_ballisticcalc.vector import Vector, ZERO_VECTOR
